Lookup config/session names of both OpenVPNCloud and CloudConnexa
Due to this rebranding, running this cockpit module on a host already
configured before the rebranding, the old config and session names in
use will be OpenVPNCloud.  With the rebranding alone, this cockpit
module would not find the the OpenVPNCloud names at all.

This adds support to be able to lookup both configurations and sessions
using the OpenVPNCloud name and manage them as if they were named
CloudConnexa.

Signed-off-by: David Sommerseth <davids@openvpn.net>